chromium/third_party/blink/web_tests/editing/pasteboard/mathml-sanitizer-bypass.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
let supportsEditableMathML =
    window.internals && internals.runtimeFlags.editingNGEnabled;

if (supportsEditableMathML) {
    selection_test(
        '<math id="target">a</math>',
        selection => {
            const target = selection.document.getElementById('target');
            const style = selection.window.getComputedStyle(target);
            supportsEditableMathML = style.display === 'inline-math';
        },
        '<math id="target">a</math>',
        'Check whether we have MathML or not');
}

// crbug.com/1011950
selection_test(
  '<div contenteditable>te|st</div>',
  selection => {
    selection.setClipboardData('<math><xss style=display:block>t<style>X<a title="</style><img src onerror=alert(1)>">.<a>.');
    selection.document.execCommand('paste');
  },
  '<div contenteditable>te<br>t<img src>">.<a>.|</a>st</div>',
  'Paste blocks script injection');

// crbug.com/1017871
selection_test(
  '<div contenteditable>te|st</div>',
  selection => {
    selection.setClipboardData('<math><xss style=display:block>t<style>X<a title="</style><style>*{background:red}</style>">.<a>.');
    selection.document.execCommand('paste');
  },
  '<div contenteditable>te<br>t">.<a>.|</a>st</div>',
  'Paste blocks HTML style injection');

// crbug.com/1017871
selection_test(
  '<div contenteditable>te|st</div>',
  selection => {
    selection.setClipboardData('A<math>B<a style=display:block>C<title>D<a id="</title><svg><style>*{background:red}</style>">c');
    selection.document.execCommand('paste');
  },
  supportsEditableMathML
    ? '<div contenteditable>teA<br>C|<svg></svg>st</div>'
    : '<div contenteditable>teA<math>B<br></math>C|<svg></svg>st</div>',
  'Paste blocks SVG style injection');
</script>
