third_party/blink/web_tests/flag-specific/disable-layout-ng/editing/pasteboard/mathml-sanitizer-bypass-expected.txt

This is a testharness.js-based test.
FAIL Paste blocks script injection resources/testharness.js:1587:25)
	 expected <div contenteditable>tet<img src>">.<a>.|</a>st</div>,
	 but got  <div contenteditable>te<br>t<img src>">.<a>.|</a>st</div>,
	 sameupto <div contenteditable>te
FAIL Paste blocks HTML style injection resources/testharness.js:1587:25)
	 expected <div contenteditable>tet">.<a>.|</a>st</div>,
	 but got  <div contenteditable>te<br>t">.<a>.|</a>st</div>,
	 sameupto <div contenteditable>te
FAIL Paste blocks SVG style injection resources/testharness.js:1587:25)
	 expected <div contenteditable>teA<math>B</math>C|<svg></svg>st</div>,
	 but got  <div contenteditable>teA<math>B<br></math>C|<svg></svg>st</div>,
	 sameupto <div contenteditable>teA<math>B<
Harness: the test ran to completion.
